char: fsm_dp: out of receive buffer handling
When out of recieve buffers, we do not replenish receive channel transfer
ring with new buffers. If the problem persists, MHI on IPA may hang.
The hang problem persists, even after out of buffer problem is resolved.

The solution is to feed the ring with a dummy buffer when out of
buffer problem happens. On the receive of such buffer, the receive packet
will be dropped.

Change-Id: I3a124bcdc160ebedc2d5c23af467a764fa331356
CRs-Fixed: 2546832
Signed-off-by: Jie Luo <jluo@codeaurora.org>